requirements.txt
ply.yacc
ply.lex